requirement.txt
adium-theme-ubuntu==0.3.4
Django==1.9.7
djangorestframework==3.3.3
dpkt==1.8.7
futures==3.0.5
psycopg2==2.6.2
pypcap==1.1.5
unity-lens-photos==1.0
virtualenv==15.0.2
